requirements.txt
datasets>=2.19.0
numpy>=1.26.0
scikit-learn>=1.4.0
torch>=2.2.0
transformers>=4.39.0
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=4.25.0
streamlit>=1.30.0
google-generativeai>=0.8.0
